+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 * Support for multiple web clients: `curl` (default) and `wget`.
 * Migrated environment variable configuration to `loda/setup.txt`.
 * Improved messages and usability of the setup command.
+* Replace `match` command by extending `mine` with an optional argument.